+++ rust/scx_utils/Cargo.toml
@@ -23,6 +23,7 @@ log = "0.4.17"
 regex = "1.10"
 sscanf = "0.4"
 tar = "0.4"
+walkdir = "2.4"
 version-compare = "0.1"
 
 [build-dependencies]
